ocr_service/requirements.txt
flask
pdfplumber
pytesseract
pillow
easyocr
requests
